# api/src/__generated__/openapi.yaml
openapi: 3.0.0
paths:
...
  /owners/{ownername}/projects/{project}/agent/suggestions/{suggestionId}/feedback:
    put:
      description: 提案を更新します
      operationId: AgentsController_updateFeedback
      parameters:
        - name: ownername
          required: true
          in: path
          description: Owner名
          schema:
            example: user1
            type: string
        - name: project
          required: true
          in: path
          description: Project名
          schema:
            example: project1
            type: string
        - name: suggestionId
          required: true
          in: path
          description: 提案のID
          schema:
            example: 123e4567-e89b-12d3-a456-426614174000
            type: string
      requestBody:
        required: true
        description: フィードバック情報
        content:
          application/json:
            schema:
              type: object
              properties:
                state:
                  type: string
                  enum:
                    - approved
                    - disapproved
                comment:
                  type: string
                reaction:
                  type: string
                  enum:
                    - good
                    - bad
              required:
                - state
      responses:
        '200':
          description: フィードバックの更新に成功しました
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    description: ステータスコード
                    type: number
                  message:
                    description: メッセージ
                    type: string
                  data:
                    type: object
                    properties:
                      suggestionId:
                        type: string
                        format: uuid
                    required:
                      - suggestionId
                required:
                  - statusCode
                  - message
                  - data
      security: &ref_8
        - supabase-auth: []
      summary: AgentsController_updateFeedback / 提案の更新
      tags: &ref_9
        - agent
  /owners/{ownername}/projects/{project}/agent/vector:
    post:
      description: vectorDBを更新します
      operationId: AgentsController_updateVector
      parameters:
        - name: ownername
          required: true
          in: path
          description: Owner名
          schema:
            example: user1
            type: string
        - name: project
          required: true
          in: path
          description: Project名
          schema:
            example: project1
            type: string
      requestBody:
        required: true
        description: トピックの更新情報
        content:
          application/json:
            schema:
              type: object
              properties:
                updateTopicIds:
                  type: array
                  items:
                    type: string
                    format: uuid
              required:
                - updateTopicIds
      responses:
        '200':
          description: vectorDBの更新に成功しました
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    description: ステータスコード
                    type: number
                  message:
                    description: メッセージ
                    type: string
                required:
                  - statusCode
                  - message
      security: *ref_8
      summary: AgentsController_updateVector / vectorDBの更新
      tags: *ref_9
  /owners/{ownername}/projects/{project}/agent/topics:
    get:
      description: トピック構築提案を発火させます
      operationId: AgentsController_triggerTopicSuggestion
      parameters:
        - name: ownername
          required: true
          in: path
          description: Owner名
          schema:
            example: user1
            type: string
        - name: project
          required: true
          in: path
          description: Project名
          schema:
            example: project1
            type: string
      responses:
        '200':
          description: トピック構築提案の発火に成功しました
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    description: ステータスコード
                    type: number
                  message:
                    description: メッセージ
                    type: string
                required:
                  - statusCode
                  - message
      security: *ref_8
      summary: AgentsController_triggerTopicSuggestion / トピック構築提案発火
      tags: *ref_9
    post:
      description: トピック構築提案の保存行います
      operationId: AgentsController_createTopicSuggestion
      parameters:
        - name: ownername
          required: true
          in: path
          description: Owner名
          schema:
            example: user1
            type: string
        - name: project
          required: true
          in: path
          description: Project名
          schema:
            example: project1
            type: string
      requestBody:
        required: true
        description: トピック構築提案の情報
        content:
          application/json:
            schema:
              type: object
              properties:
                metadata:
                  type: object
                  properties:
                    project_id:
                      type: string
                      format: uuid
                    agent_id:
                      type: string
                      format: uuid
                  required:
                    - project_id
                    - agent_id
                suggestion:
                  type: object
                  properties:
                    id:
                      type: string
                      format: uuid
                    type:
                      type: string
                    title:
                      type: string
                    accuracy:
                      type: number
                  required:
                    - id
                    - type
                    - title
                    - accuracy
                data:
                  type: object
                  properties:
                    id:
                      type: string
                      format: uuid
                    frame_id:
                      type: string
                      format: uuid
                    title:
                      type: string
                    state:
                      type: string
                      enum:
                        - open
                        - closed
                    assignee:
                      type: string
                      format: uuid
                    reviewer:
                      type: string
                      format: uuid
                    participants:
                      type: array
                      items:
                        type: string
                        format: uuid
                    tags:
                      type: array
                      items:
                        type: object
                        properties:
                          name:
                            type: string
                          color:
                            type: string
                        required:
                          - name
                          - color
                    attributes:
                      type: array
                      items:
                        type: object
                        properties:
                          name:
                            type: string
                          content:
                            oneOf:
                              - oneOf:
                                  - type: string
                                  - type: number
                                  - type: boolean
                                  - {}
                              - type: array
                                items: {}
                              - type: object
                                additionalProperties: {}
                        required:
                          - name
                          - content
                    document: {}
                    contexts:
                      type: array
                      items:
                        type: object
                        properties:
                          id:
                            type: string
                            format: uuid
                          name:
                            type: string
                          description:
                            type: string
                          type:
                            type: string
                          opponent_topic_id:
                            type: string
                            format: uuid
                        required:
                          - name
                          - description
                          - type
                          - opponent_topic_id
                  required:
                    - frame_id
                    - title
                    - state
                    - assignee
                    - reviewer
                    - participants
                    - tags
                    - attributes
                    - document
                    - contexts
              required:
                - metadata
                - suggestion
                - data
            examples:
              default:
                value:
                  metadata:
                    project_id: 3fa85f64-5717-4562-b3fc-2c963f66afa6
                    agent_id: 3fa85f64-5717-4562-b3fc-2c963f66afa6
                  suggestion:
                    id: 3fa85f64-5717-4562-b3fc-2c963f66afa6
                    type: A
                    title: これはタイトル
                    accuracy: 0.95
                  data:
                    id: 3fa85f64-5717-4562-b3fc-2c963f66afa6
                    frame_id: 3fa85f64-5717-4562-b3fc-2c963f66afa6
                    title: これはタイトル
                    state: open
                    assignee: 3fa85f64-5717-4562-b3fc-2c963f66afa6
                    reviewer: 3fa85f64-5717-4562-b3fc-2c963f66afa6
                    participants:
                      - 3fa85f64-5717-4562-b3fc-2c963f66afa6
                    tags:
                      - name: tag1
                        color: '#ff0000'
                      - name: tag2
                        color: '#00ff00'
                    attributes:
                      - name: attribute1
                        content: content1
                      - name: attribute2
                        content: content2
                    document:
                      type: doc
                      content: []
                    contexts:
                      - id: 3fa85f64-5717-4562-b3fc-2c963f66afa6
                        name: context1
                        description: description1
                        type: type1
                        opponent_topic_id: 3fa85f64-5717-4562-b3fc-2c963f66afa6
                      - id: 3fa85f64-5717-4562-b3fc-2c963f66afa6
                        name: context2
                        description: description2
                        type: type2
                        opponent_topic_id: 3fa85f64-5717-4562-b3fc-2c963f66afa6
      responses:
        '200':
          description: トピック構築提案の作成に成功しました
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    description: ステータスコード
                    type: number
                  message:
                    description: メッセージ
                    type: string
                  data:
                    type: object
                    properties:
                      topicId:
                        type: string
                        format: uuid
                      suggestionId:
                        type: string
                        format: uuid
                    required:
                      - topicId
                      - suggestionId
                required:
                  - statusCode
                  - message
                  - data
      security: *ref_8
      summary: AgentsController_createTopicSuggestion / トピック構築提案保存
      tags: *ref_9
  /owners/{ownername}/projects/{project}/agent/topics/{topicId}/timeline:
    get:
      description: 検討もれ提案を発火させます
      operationId: AgentsController_triggerChatSuggestion
      parameters:
        - name: ownername
          required: true
          in: path
          description: Owner名
          schema:
            example: user1
            type: string
        - name: project
          required: true
          in: path
          description: Project名
          schema:
            example: project1
            type: string
        - name: topicId
          required: true
          in: path
          description: トピックのID
          schema:
            example: 123e4567-e89b-12d3-a456-426614174000
            type: string
      responses:
        '200':
          description: 検討もれ提案の発火に成功しました
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    description: ステータスコード
                    type: number
                  message:
                    description: メッセージ
                    type: string
                required:
                  - statusCode
                  - message
      security: *ref_8
      summary: AgentsController_triggerChatSuggestion / 検討もれ提案の発火
      tags: *ref_9
    post:
      description: 検討もれ提案の保存行います
      operationId: AgentsController_createChatSuggestion
      parameters:
        - name: ownername
          required: true
          in: path
          description: Owner名
          schema:
            example: user1
            type: string
        - name: project
          required: true
          in: path
          description: Project名
          schema:
            example: project1
            type: string
        - name: topicId
          required: true
          in: path
          description: トピックのID
          schema:
            example: 123e4567-e89b-12d3-a456-426614174000
            type: string
      requestBody:
        required: true
        description: タイムライン内提案情報
        content:
          application/json:
            schema:
              type: object
              properties:
                metadata:
                  type: object
                  properties:
                    request_type:
                      type: string
                    project_id:
                      type: string
                      format: uuid
                    agent_id:
                      type: string
                      format: uuid
                    activity_id:
                      type: string
                      format: uuid
                  required:
                    - request_type
                    - project_id
                    - agent_id
                    - activity_id
                suggestion:
                  type: object
                  properties:
                    id:
                      type: string
                      format: uuid
                    type:
                      type: string
                    title:
                      type: string
                    accuracy:
                      type: number
                  required:
                    - type
                    - title
                    - accuracy
                data:
                  oneOf:
                    - oneOf:
                        - type: string
                        - type: number
                        - type: boolean
                        - {}
                    - type: array
                      items: {}
                    - type: object
                      additionalProperties: {}
              required:
                - metadata
                - suggestion
                - data
            examples:
              default:
                value:
                  metadata:
                    request_type: string
                    project_id: 3fa85f64-5717-4562-b3fc-2c963f66afa6
                    agent_id: 3fa85f64-5717-4562-b3fc-2c963f66afa6
                    activity_id: 3fa85f64-5717-4562-b3fc-2c963f66afa6
                  suggestion:
                    id: 3fa85f64-5717-4562-b3fc-2c963f66afa6
                    type: string
                    title: string
                    accuracy: 0
                  data:
                    id: 3fa85f64-5717-4562-b3fc-2c963f66afa6
                    prop1: string
      responses:
        '200':
          description: 検討もれ提案の作成に成功しました
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    description: ステータスコード
                    type: number
                  message:
                    description: メッセージ
                    type: string
                  data:
                    type: object
                    properties:
                      activityId:
                        type: string
                        format: uuid
                      suggestionId:
                        type: string
                        format: uuid
                    required:
                      - activityId
                      - suggestionId
                required:
                  - statusCode
                  - message
                  - data
      security: *ref_8
      summary: AgentsController_createChatSuggestion / 検討もれ提案保存
      tags: *ref_9
info:
  title: andTopic API
  description: andTopic API description
  version: '1.0'
  contact: {}
...
servers: []
...